change float to double

+++ ServiceExample/NumberService.cs
@@ -10,12 +10,12 @@ namespace ServiceExample
     {
         public string ReturnResults(string stringNum1, string stringNum2)
         {
-            float num1;
-            float num2;
+            double num1;
+            double num2;
 
             try
             {
-                num1 = float.Parse(stringNum1);
+                num1 = Double.Parse(stringNum1);
             }
             catch
             {
@@ -24,17 +24,17 @@ namespace ServiceExample
 
             try
             {
-                num2 = float.Parse(stringNum2);
+                num2 = Double.Parse(stringNum2);
             }
             catch
             {
                 return "The second input is not a number, please try again.";
             }
 
-            float numSum = num1 + num2;
-            float numDiff = num1 - num2;
-            float numProd = num1 * num2;
-            float numFrac = num1 / num2;
+            double numSum = num1 + num2;
+            double numDiff = num1 - num2;
+            double numProd = num1 * num2;
+            double numFrac = num1 / num2;
 
             return $"The sum of those two numbers is: {numSum}\n" +
                 $"The difference between those two numbers is: {numDiff}\n" +
